refactor: oRPC contract-first architecture
Separates contracts (client-safe) from implementations (server-only) to prevent
node:fs, better-sqlite3, etc. from being pulled into the browser bundle.

- Add src/contracts/ with client-safe contract definitions
- Move src/lib/orpc.ts to src/server/orpc.ts with implement(contract)
- Update client to import contract instead of router
- Update server modules to use os.<module>.<procedure>.handler() pattern
- Change hooks to use ContractInputs instead of RouterInputs
