+++ apps/app/src/lib/orpc-client.ts
@@ -1,13 +1,20 @@
 import { createORPCClient } from "@orpc/client";
-import { RPCLink } from "@orpc/client/fetch";
+import type { ContractRouterClient } from "@orpc/contract";
+import type { JsonifiedClient } from "@orpc/openapi-client";
+import { OpenAPILink } from "@orpc/openapi-client/fetch";
 import { createORPCReactQueryUtils } from "@orpc/react-query";
-import type { RouterClient } from "@orpc/server";
-import type { Router } from "@/server/index";
+import { type Contract, contract } from "@/contracts";
 
-const link = new RPCLink({
-  url: "/api/rpc",
+const link = new OpenAPILink(contract, {
+  url: () => {
+    if (typeof window !== "undefined") {
+      return `${window.location.origin}/api`;
+    }
+    return "http://localhost:3000/api";
+  },
 });
 
-export const client: RouterClient<Router> = createORPCClient(link);
+export const client: JsonifiedClient<ContractRouterClient<Contract>> =
+  createORPCClient(link);
 
 export const orpc = createORPCReactQueryUtils(client);
