+++ apps/app/src/server/index.ts
@@ -1,14 +1,14 @@
-import type { InferRouterInputs, InferRouterOutputs } from "@orpc/server";
 import { apiKeys } from "./api-keys";
 import { deployments } from "./deployments";
 import { domains } from "./domains";
 import { health } from "./health";
+import { os } from "./orpc";
 import { projects } from "./projects";
 import { registries } from "./registries";
 import { services } from "./services";
 import { dbTemplates, templates } from "./templates";
 
-export const router = {
+export const router = os.router({
   apiKeys,
   dbTemplates,
   deployments,
@@ -18,8 +18,6 @@ export const router = {
   registries,
   services,
   templates,
-};
+});
 
 export type Router = typeof router;
-export type RouterInputs = InferRouterInputs<Router>;
-export type RouterOutputs = InferRouterOutputs<Router>;
